Fix debug build for MSVC2015
